Introduced visualisation module. Moved visualisation code out of the alignment tree module. Grouped imports in reptilian.py to prevent warnings (imports should be at the top of the file).
